feat:refactor into smaller components keeping codes simple and clean
